Laravel Project
Web application for a school admin to generate dynamic timetable for schools(faculty) based on number of subjects provided. The system allows the user to do authentication, add lecturers, subjects, and generate the timetable accordingly. It was built using PHP on Laravel framework. MVC structure was applied while developing the application.